(My third commit)

+++ week-3/practice_9/src/main.rs
@@ -1,5 +1,5 @@
 fn main() {
-    let mut fees:i32 = 25_000;
+    let mut fees:u16 = 25_000;
     println!("fees is {}", fees);
 
     fees = 35_000;
